\name{hd2amelia}
\alias{hd2amelia}
\title{Convert hot.deck output to Amelia format
}
\description{
Converts the output from hot.deck to the format used by Amelia for use with the Zelig package. 
}
\usage{
hd2amelia(object)
}
\arguments{
  \item{object}{Output from a run of the \code{hot.deck} function. }
}
\value{
An object of class \dQuote{amelia} that can be used with Zelig.}
\author{
Skyler Cranmer, Jeff Gill, Natalie Jackson, Andreas Murr and Dave Armstrong
}
