\name{isq99}
\alias{isq99}
\docType{data}
\title{Data from Poe, Tate and Keith 1999.}
\usage{data(isq99)}
\description{Data on Democracy, State Repression and other state-level characteristics}

\format{
  A data frame with 3222 observations on the following 13 variables.
  \describe{
    \item{\code{IDORIGIN}}{Country Code}
    \item{\code{YEAR}}{Year}
    \item{\code{AI}}{Amnesty International PTS Coding}
    \item{\code{SD}}{State Department Country Report PTS Coding}
    \item{\code{POLRT}}{Freedom House Political Rights Variable}
    \item{\code{MIL2}}{Military Government}
    \item{\code{LEFT}}{Leftist Government}
    \item{\code{BRIT}}{British Colonial Influence}
    \item{\code{PCGNP}}{GNP/capita}
    \item{\code{LPOP}}{Log of population}
    \item{\code{DEMOC3}}{Polity III Democracy}
    \item{\code{CWARCOW}}{COW Civil War}
    \item{\code{IWARCOW2}}{COW Interstate War}
  }
}
\references{
Steven Poe, C. Neal Tate and Linda Camp Keith.  1999.  \dQuote{Repression of the Human Right to Personal Integrity Revisited: A Global, Cross-National Study Covering the Years 1976-1993}.  International Studies Quarterly.  43: 291-313.
}
\keyword{datasets}


