% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd2amelia.R
\name{hd2amelia}
\alias{hd2amelia}
\title{Convert hot.deck output to Amelia format.}
\usage{
hd2amelia(object)
}
\arguments{
\item{object}{Output from a run of the \code{hot.deck} function.}
}
\value{
An object of class \dQuote{amelia} that can be used with Zelig.
}
\description{
Converts the output from hot.deck to the format used by Amelia for use with the Zelig package.
}
